+++ qb_move_hardware_interface/CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog for package qb_move_hardware_interface
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.1.0 (2019-05-28)
+------------------
+* Fix minor style issues
+* Add an interactive marker to control the device
+* Fix minor style issues
+
 2.0.2 (2018-08-09)
 ------------------
 * Update license agreement copyright